requirements_clean.txt
absl-py
accelerate==1.3.0
aiohappyeyeballs
aiohttp
aiohttp-cors==0.7.0
aiosignal
alembic
annotated-types
ansi2html
anyio
argon2-cffi
argon2-cffi-bindings
array-record==0.4.1
arrow
arviz
astroid
astropy
astropy-iers-data
asttokens
astunparse
async-lru
async-timeout
attrs
autofeat==2.1.3
autograd==1.7.0
autograd-gamma==0.5.0
autopep8==2.3.2
babel
backcall
bandit==1.8.0
bayesian-optimization==1.4.3
bcrypt
beautifulsoup4
black==25.1.0
bleach
blinker
blis
bokeh
Bottleneck
Brotli
build==1.2.2.post1
cached-property
cachetools
catalogue
catboost 
category_encoders
certifi
cffi
cftime
charset-normalizer
chart-studio
click
cloudpathlib
cloudpickle
colorama
colorful==0.5.6
colorlog
colorlover==0.3.0
comm
confection
confuse
cons
contourpy
cramjam
cryptography
cufflinks
cycler
cymem==2.0.10
Cython
cytoolz
dash
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask
dask-expr
dask-glm
dask-jobqueue
dask-ml
databricks-sdk
dataclasses
datasets
datefinder
debugpy
decorator
defusedxml
deprecat
Deprecated
deprecation
dill
distlib==0.3.9
distributed
dm-tree==0.1.8
dnspython
docker
docker-pycreds==0.4.0
docutils==0.21.2
dodgy
email_validator
entrypoints
et_xmlfile
etils==1.5.2
etuples
eval_type_backport==0.2.2
exceptiongroup
executing
fake-useragent
fastapi
fastapi-cli
fastcore==1.7.23
fastdownload==0.0.7
fastjsonschema
fastparquet
fastprogress
featuretools
filelock
flake8
flake8-polyfill
Flask==3.0.3
Flask-Cors
flatbuffers
flexcache==0.3
flexparser==0.4
fonttools
formulaic==1.1.1
fqdn
frozenlist
fsspec
fst-pso
funcy
future
FuzzyTM
gast
gensim
gitdb
GitPython
google-api-core==2.24.1
google-auth
google-auth-oauthlib==1.0.0
google-pasta
googleapis-common-protos==1.67.0
graphene
graphql-core
graphql-relay
graphviz
greenlet
h11
h2
h2o==3.46.0.6
h5netcdf
h5py==3.12.1
holidays
hpack
htmlmin
http_client==0.1.22
httpcore
httptools
httpx
huggingface_hub
hyperframe
icecream
idna
ImageHash
imbalanced-learn
importlib_metadata
importlib_resources
iniconfig==2.0.0
interface-meta==1.3.0
ipykernel
ipython
ipywidgets
isoduration
isort
itsdangerous
jax==0.4.30
jaxlib==0.4.30
jedi
Jinja2
joblib
json5
JsonForm==0.0.2
jsonpickle==4.0.1
jsonpointer
jsonschema
jsonschema-specifications
JsonSir==0.0.2
jupyter
jupyter-console
jupyter-dash
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kaleido
keras==3.8.0
Keras_Preprocessing
kiwisolver
kmodes
langcodes
language_data
libclang==18.1.1
lightgbm
llvmlite==0.42.0
locket
logical-unification
logilab-common
lttbc
lxml==5.3.0
lz4
Mako
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.7.5
matplotlib-inline
mccabe
mdurl
memory_profiler
miniful
miniKanren
missingno
mistune
mkl-service==2.4.1
mkl_fft==1.3.10
mkl_random
ml-dtypes==0.3.2
mlflow==2.19.0
mlflow-skinny
mlxtend
mpmath
msgpack
multidict
multimethod
multipledispatch
multiprocess
munkres==1.1.4
murmurhash
mypy==1.14.1
mypy_extensions
namex==0.0.8
nbclient
nbconvert
nbformat
nest_asyncio
netCDF4
networkx
ngboost==0.5.2
nltk
notebook
notebook_shim
numba
numexpr
numpy==1.23.5
oauthlib
opencensus==0.11.4
opencensus-context==0.1.3
openpyxl
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
opt_einsum
optree==0.13.1
optuna==4.2.1
orjson
outcome
overrides
packaging
pandas
pandas-profiling
pandocfilters
paramiko
parso
partd
pathos==0.3.3
pathspec==0.12.1
patsy
pbr==6.1.0
pep8-naming==0.10.0
pgmpy==0.1.19
phik
pickleshare
Pillow
Pint==0.24.4
pkgutil_resolve_name
platformdirs
plotly-resampler==0.10.0
pluggy==1.5.0
ply
pmdarima
polars
pox==0.3.5
ppft==1.7.6.9
preshed
prometheus_client
prometheus_flask_exporter
promise==2.3
prompt_toolkit
propcache
prospector
proto-plus==1.26.0
protobuf
psutil
pure_eval
py-spy==0.4.0
pyarrow==14.0.1
pyarrow-hotfix
pyasn1
pyasn1_modules
pycodestyle
pycparser
pycurl==7.45.4
pydantic
pydantic_core
pydocstyle
pyerfa
pyflakes
pyFUME
Pygments
pygpu==0.7.6
PyJWT
pyLDAvis
pylint
pylint-celery==0.3
pylint-django
pylint-flask==0.6
pylint-plugin-utils
pymc
pymc3
pymongo
PyNaCl
pynndescent
pyod
pyOpenSSL
pyparsing
pyproject_hooks==1.2.0
PyQt5==5.15.10
PyQt5-sip
pyroma==4.2
PySnooper
PySocks
pytensor
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil
python-dotenv
Python-EasyConfig==0.1.7
python-json-logger
python-multipart
pytz
pyu2f
pyvis==0.3.2
PyWavelets==1.6.0
pywin32==307
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
querystring_parser
RapidFuzz==3.10.1
ray==2.42.1
referencing
regex
requests
requests-oauthlib
requirements-detector
Resource==0.2.1
responses
retrying
rfc3339_validator
rfc3986-validator
rich
rich-toolkit
rpds-py
rsa
safetensors
schemdraw
scikit-base
scikit-learn
scikit-plot==0.3.7
SciPy
selenium
semver
Send2Trash
sentry-sdk==2.19.2
setoptconf-tmp
setproctitle
shap
shellingham
simpful
sip
six
slicer
smart-open==7.0.5
smmap==4.0.0
sniffio
snowballstemmer
sortedcontainers
soupsieve
spacy
spacy-legacy
spacy-loggers
sparse
SQLAlchemy
sqlparse
srsly==2.4.8
stack_data
starlette
statsmodels
stevedore==5.4.0
structlog
sympy==1.13.1
tabulate==0.9.0
tangled-up-in-unicode
tbats
tblib
tenacity
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit
tensorboardX==2.6.2.2
tensorflow==2.16.1
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
termcolor
terminado
textblob==0.15.3
tf-keras==2.15.0
Theano
Theano-PyMC
thinc
threadpoolctl
tinycss2
tokenizers==0.21.0
toml
tomli
tomlkit
toolz
torch==2.6.0
torch-geometric==2.6.1
torchaudio==2.6.0
torchvision==0.21.0
tornado
tqdm
trace_updater
traitlets
transformers
trio
trio-websocket
trove-classifiers==2025.1.7.14
tsdownsample==0.1.3
typer==0.15.1
typer-slim==0.15.1
types-python-dateutil
typing_extensions
typing_utils
tzdata
umap-learn
unicodedata2
uri-template
urllib3
uv==0.5.29
uvicorn
virtualenv==20.29.2
visions
vulture==2.14
waitress
wasabi
watchfiles
wcwidth
weasel
webcolors
webdriver_manager
webencodings
websocket-client
websockets
Werkzeug==3.0.6
widgetsnbextension
win_inet_pton
woodwork
wordcloud
wrapt==1.14.1
wsproto
wurlitzer
xarray
xarray-einstats
xgboost
xxhash
xyzservices
yarl
yellowbrick
zict
zipp
zstandard==0.23.0
